submission/supplementary/code/core/basis.py
"""Fractal basis implementation for quantum state transformations.

From appendix_j_math_details.tex:
The fractal basis provides a complete orthonormal system for
quantum states through recursive dimensional reduction.
"""

from typing import Dict, Optional, Union, List, Tuple, Any
import numpy as np
from math import log, factorial

# Third party imports (following Sacred Import Hierarchy)
from scipy import special, integrate
from sympy import (
    Symbol, exp, integrate as sym_integrate, conjugate, sqrt,
    oo, I, pi, Matrix, diff, solve, Eq, Function,
    factorial as sym_factorial, hermite
)

# Local imports (following Sacred Constants Organization Law)
from .physics_constants import (
    HBAR, C, I,  # Level 1: Fundamental constants
    g_μν, Gamma,  # Level 2: Mathematical objects
    X, T  # Level 3: Derived quantities
)
from .types import (
    Energy, WaveFunction, NumericValue,
    FractalMode, BasisConfig
)
from .errors import (
    BasisError, ValidationError, ComputationError,
    CoherenceError
)

class FractalBasis:
    """
    Implementation of fractal basis expansion.
    
    From appendix_b_basis.tex:
    The fractal basis provides a natural decomposition of
    quantum states through recursive scaling transformations.
    """
    
    def __init__(
        self,
        alpha: float = 0.1,
        precision: float = 1e-10,
        *,
        max_level: int = 10,
        dimension: int = 4
    ):
        """Initialize fractal basis."""
        self.alpha = alpha
        self.precision = precision
        self.max_level = max_level
        self.dimension = dimension
        self.scaling_dimension = (dimension - 2)/2 

    def compute_basis_functions(self, config: BasisConfig) -> List[FractalMode]:
        """
        Compute fractal basis functions.
        
        From appendix_b_basis.tex Eq B.1-B.3:
        The basis functions are constructed through recursive scaling:
        ψ_n(x) = α^{n/2} ψ₀(α^n x)
        
        where ψ₀ is the ground state and α is the scaling parameter.
        """
        try:
            # SACRED: Use simple complex type
            grid = np.linspace(-3, 3, 100)  # SACRED: grid range
            
            # Compute basis functions for each level
            basis_functions = []
            for n in range(self.max_level):
                # SACRED: Preserve coordinate scaling
                x_scaled = grid * self.alpha**n
                
                # Compute mode function with proper normalization
                mode = self._compute_mode_function(n, x_scaled, config)
                
                # Create basis function with quantum numbers
                basis_functions.append(FractalMode(
                    psi=mode,
                    grid=grid,
                    n=n,
                    alpha=self.alpha
                ))
                
            return basis_functions
            
        except Exception as e:
            raise BasisError(f"Basis function computation failed: {e}")
            
    def _compute_mode_function(self, n: int, x_scaled: np.ndarray, 
                             config: BasisConfig) -> np.ndarray:
        """
        Compute nth mode function.
        
        From appendix_b_basis.tex Eq B.4:
        The mode functions include both oscillatory and
        damping terms to ensure proper convergence.
        """
        # SACRED: Use simple complex type
        try:
            # Compute oscillatory part
            k = 2*pi*n  # Wavevector
            oscillation = np.exp(I * k * x_scaled)
            
            # Compute damping envelope
            damping = np.exp(-x_scaled**2/(2*HBAR))
            
            # SACRED: Two-step normalization
            # First compute raw mode
            mode = oscillation * damping
            
            # Then normalize
            dx = x_scaled[1] - x_scaled[0]
            norm = 1/np.sqrt(np.sum(np.abs(mode)**2) * dx)
            
            return norm * mode
            
        except Exception as e:
            raise BasisError(f"Mode function computation failed: {e}")

    def project_state(self, psi: WaveFunction) -> Dict[int, complex]:
        """
        Project quantum state onto fractal basis.
        
        From appendix_b_basis.tex Eq B.7-B.9:
        The projection coefficients are computed through:
        c_n = ∫dx ψ_n*(x)ψ(x)
        """
        try:
            # Validate input state
            if not isinstance(psi, WaveFunction):
                raise ValidationError("Input must be a WaveFunction")
                
            # Compute basis functions
            basis = self.compute_basis_functions(BasisConfig(
                dimension=self.dimension,
                precision=self.precision
            ))
            
            # Project onto each basis function
            coefficients = {}
            for mode in basis:
                # SACRED: Preserve coordinate scaling
                coeff = self._compute_projection(psi, mode)
                
                # Store if above precision threshold
                if abs(coeff) > self.precision:
                    coefficients[mode.n] = coeff
                    
            return coefficients
            
        except Exception as e:
            raise BasisError(f"State projection failed: {e}")

    def _compute_projection(self, psi: WaveFunction, mode: FractalMode) -> complex:
        """
        Compute projection coefficient onto basis mode.
        
        From appendix_b_basis.tex Eq B.10:
        The projection preserves orthonormality:
        ∫dx ψ_m*(x)ψ_n(x) = δ_mn
        """
        try:
            # SACRED: Use simple complex type
            dx = mode.grid[1] - mode.grid[0]
            
            # Compute overlap integral
            overlap = np.sum(
                np.conjugate(mode.psi) * psi.psi
            ) * dx
            
            # Validate against wavelet_analysis.csv thresholds
            self._validate_projection(overlap, mode.n)
            
            return complex(overlap)
            
        except Exception as e:
            raise BasisError(f"Projection computation failed: {e}")

    def _validate_projection(self, overlap: complex, n: int) -> None:
        """
        Validate projection coefficient.
        
        From appendix_b_basis.tex Eq B.11-B.13:
        The coefficients must satisfy:
        1. Normalization: ∑|c_n|² = 1
        2. Scaling: |c_n| ~ α^{nΔ}
        3. Phase coherence
        """
        try:
            # Load validation thresholds
            thresholds = self._load_wavelet_thresholds()
            
            # Check normalization
            if abs(overlap) > 1.0:
                raise CoherenceError("Projection coefficient exceeds unity")
                
            # Verify scaling behavior
            expected_scaling = self.alpha**(n * self.scaling_dimension)
            if abs(abs(overlap) - expected_scaling) > thresholds['scaling']:
                raise CoherenceError("Invalid scaling behavior")
                
            # Check phase coherence
            phase = np.angle(overlap)
            if not self._check_phase_coherence(phase, n):
                raise CoherenceError("Phase coherence violated")
                
        except Exception as e:
            raise ValidationError(f"Projection validation failed: {e}")

    def _check_phase_coherence(self, phase: float, n: int) -> bool:
        """
        Check phase coherence of projection.
        
        From appendix_b_basis.tex Eq B.14:
        The phases must follow the fractal pattern:
        φ_n = 2πn/N + α^n θ
        """
        # Load phase data
        coherence_data = self._load_wavelet_analysis()
        
        # Compute expected phase
        expected = 2*pi*n/self.max_level + self.alpha**n * coherence_data['theta']
        
        # Check against threshold
        return abs(phase - expected) < coherence_data['phase_threshold']

    def _load_wavelet_thresholds(self) -> Dict[str, float]:
        """Load wavelet analysis thresholds."""
        try:
            thresholds = {}
            with open('supplementary/data/wavelet_analysis.csv', 'r') as f:
                for line in f:
                    if line.startswith('#'):
                        continue
                    name, value = line.strip().split(',')
                    thresholds[name] = float(value)
            return thresholds
            
        except Exception as e:
            raise ValidationError(f"Failed to load wavelet thresholds: {e}")

    def analyze_fractal_scaling(self, psi: WaveFunction) -> Dict[str, NumericValue]:
        """
        Analyze fractal scaling properties of quantum state.
        
        From appendix_b_basis.tex Eq B.15-B.17:
        The fractal dimension is computed through:
        D_f = lim_{n→∞} log N(n)/log(1/α^n)
        
        where N(n) is the number of significant coefficients.
        """
        try:
            # Project state onto basis
            coefficients = self.project_state(psi)
            
            # SACRED: Two-step analysis
            # First compute scaling exponents
            exponents = self._compute_scaling_exponents(coefficients)
            
            # Then analyze fractal properties
            results = {
                'fractal_dimension': self._compute_fractal_dimension(exponents),
                'scaling_symmetry': self._verify_scaling_symmetry(exponents),
                'coherence_length': self._compute_coherence_length(coefficients)
            }
            
            # Validate against statistical_analysis.csv
            self._validate_fractal_analysis(results)
            
            return results
            
        except Exception as e:
            raise BasisError(f"Fractal analysis failed: {e}")

    def _compute_scaling_exponents(self, coeffs: Dict[int, complex]) -> np.ndarray:
        """
        Compute scaling exponents from coefficients.
        
        From appendix_b_basis.tex Eq B.18:
        The scaling exponents follow:
        γ(n) = log|c_n|/log(α^n)
        """
        try:
            exponents = []
            for n, c_n in coeffs.items():
                # Skip zero coefficients
                if abs(c_n) < self.precision:
                    continue
                    
                # Compute scaling exponent
                gamma = log(abs(c_n))/log(self.alpha**n)
                exponents.append(gamma)
                
            return np.array(exponents)
            
        except Exception as e:
            raise ComputationError(f"Scaling exponent computation failed: {e}")

    def _compute_fractal_dimension(self, exponents: np.ndarray) -> NumericValue:
        """
        Compute fractal dimension from scaling exponents.
        
        From appendix_b_basis.tex Eq B.19:
        D_f = 2 - lim_{n→∞} γ(n)
        """
        try:
            # Load statistical thresholds
            stats = self._load_statistical_analysis()
            
            # Compute dimension with uncertainty
            dim = 2 - np.mean(exponents)
            uncertainty = np.std(exponents) / sqrt(len(exponents))
            
            # Validate against expected dimension
            expected = stats['expected_dimension']
            if abs(dim - expected) > 3*uncertainty:
                raise ValidationError(
                    f"Fractal dimension {dim:.3f} ± {uncertainty:.3f} "
                    f"deviates from expected {expected:.3f}"
                )
                
            return NumericValue(dim, uncertainty)
            
        except Exception as e:
            raise ComputationError(f"Fractal dimension computation failed: {e}")

    def _verify_scaling_symmetry(self, exponents: np.ndarray) -> NumericValue:
        """
        Verify scaling symmetry of basis expansion.
        
        From appendix_b_basis.tex Eq B.20:
        The scaling symmetry requires:
        γ(n+m) = γ(n) + γ(m) + O(α^max(n,m))
        """
        try:
            # Compute symmetry violations
            violations = []
            for i in range(len(exponents)):
                for j in range(i+1, len(exponents)):
                    violation = abs(
                        exponents[i+j] - (exponents[i] + exponents[j])
                    )
                    violations.append(violation)
                    
            # Compute mean violation and uncertainty
            mean_violation = np.mean(violations)
            uncertainty = np.std(violations) / sqrt(len(violations))
            
            return NumericValue(mean_violation, uncertainty)
            
        except Exception as e:
            raise ValidationError(f"Scaling symmetry verification failed: {e}")

    def _compute_coherence_length(self, coeffs: Dict[int, complex]) -> NumericValue:
        """
        Compute quantum coherence length from basis coefficients.
        
        From appendix_b_basis.tex Eq B.21-B.23:
        The coherence length is defined through:
        ξ = 1/α^n_c
        
        where n_c is the critical level where correlations decay.
        """
        try:
            # Get coefficient magnitudes
            magnitudes = np.array([abs(c) for c in coeffs.values()])
            levels = np.array(list(coeffs.keys()))
            
            # Find critical level where correlations decay
            correlations = self._compute_correlation_function(magnitudes)
            n_c = self._find_critical_level(correlations)
            
            # Compute coherence length with uncertainty
            xi = 1/self.alpha**n_c
            uncertainty = abs(xi * log(self.alpha))  # Error propagation
            
            return NumericValue(xi, uncertainty)
            
        except Exception as e:
            raise ComputationError(f"Coherence length computation failed: {e}")

    def _compute_correlation_function(self, magnitudes: np.ndarray) -> np.ndarray:
        """
        Compute correlation function between basis coefficients.
        
        From appendix_b_basis.tex Eq B.24:
        The correlation function measures quantum coherence:
        C(r) = <c_n* c_{n+r}> / <|c_n|²>
        """
        try:
            # Normalize magnitudes
            norm_magnitudes = magnitudes / np.sqrt(np.sum(magnitudes**2))
            
            # Compute correlations at different separations
            correlations = []
            for r in range(len(magnitudes)//2):
                corr = np.sum(norm_magnitudes[:-r] * norm_magnitudes[r:])
                correlations.append(corr)
                
            return np.array(correlations)
            
        except Exception as e:
            raise ComputationError(f"Correlation computation failed: {e}")

    def _find_critical_level(self, correlations: np.ndarray) -> int:
        """
        Find critical level where correlations decay.
        
        From appendix_b_basis.tex Eq B.25:
        The critical level is determined by:
        C(n_c) = C_0 exp(-n_c/ξ)
        """
        try:
            # Load statistical thresholds
            stats = self._load_statistical_analysis()
            threshold = stats['correlation_threshold']
            
            # Find where correlations drop below threshold
            mask = correlations < threshold
            if not np.any(mask):
                raise ValidationError("Correlations never decay below threshold")
                
            return int(np.argmax(mask))
            
        except Exception as e:
            raise ComputationError(f"Critical level determination failed: {e}")

    def _validate_fractal_analysis(self, results: Dict[str, NumericValue]) -> None:
        """
        Validate fractal analysis results.
        
        From appendix_b_basis.tex Eq B.26-B.28:
        The results must satisfy:
        1. Fractal dimension bounds
        2. Scaling symmetry
        3. Coherence length scaling
        """
        try:
            # Load validation data
            stats = self._load_statistical_analysis()
            
            # Check fractal dimension
            D_f = results['fractal_dimension'].value
            if not (1.0 < D_f < 2.0):
                raise ValidationError(f"Invalid fractal dimension: {D_f}")
                
            # Verify scaling symmetry
            sym = results['scaling_symmetry'].value
            if sym > stats['symmetry_threshold']:
                raise ValidationError(f"Scaling symmetry violated: {sym}")
                
            # Check coherence length
            xi = results['coherence_length'].value
            if xi < stats['min_coherence_length']:
                raise ValidationError(f"Coherence length too short: {xi}")
                
        except Exception as e:
            raise ValidationError(f"Fractal analysis validation failed: {e}")

    def analyze_wavelet_transform(self, psi: WaveFunction) -> Dict[str, NumericValue]:
        """
        Analyze wavelet transform properties of basis decomposition.
        
        From appendix_b_basis.tex Eq B.29-B.31:
        The wavelet analysis verifies:
        1. Localization in position and momentum
        2. Resolution of unity
        3. Admissibility condition
        """
        try:
            # Project state onto basis
            coeffs = self.project_state(psi)
            
            # SACRED: Two-step wavelet analysis
            # First compute wavelet transforms
            transforms = self._compute_wavelet_transforms(coeffs)
            
            # Then analyze wavelet properties
            results = {
                'localization': self._verify_localization(transforms),
                'resolution': self._verify_resolution(transforms),
                'admissibility': self._verify_admissibility(transforms)
            }
            
            # Validate against wavelet_analysis.csv
            self._validate_wavelet_analysis(results)
            
            return results
            
        except Exception as e:
            raise BasisError(f"Wavelet analysis failed: {e}")

    def _compute_wavelet_transforms(self, coeffs: Dict[int, complex]) -> np.ndarray:
        """
        Compute wavelet transforms of basis coefficients.
        
        From appendix_b_basis.tex Eq B.32:
        The wavelet transform is:
        W_ψ[f](a,b) = |a|^{-1/2} ∫dx f(x)ψ*((x-b)/a)
        """
        try:
            # Get coefficient array
            levels = np.array(list(coeffs.keys()))
            values = np.array(list(coeffs.values()))
            
            # Compute transforms at each scale
            transforms = []
            for a in self.alpha**levels:
                # SACRED: Preserve coordinate scaling
                x_scaled = np.linspace(-3, 3, 100)/a  # SACRED: grid range
                
                # Compute mother wavelet
                psi = np.exp(-x_scaled**2/2) * np.exp(2j*pi*x_scaled)
                
                # Compute transform
                W = np.sum(values[:, None] * np.conjugate(psi), axis=0)
                transforms.append(W/np.sqrt(abs(a)))
                
            return np.array(transforms)
            
        except Exception as e:
            raise ComputationError(f"Wavelet transform computation failed: {e}")

    def _verify_localization(self, transforms: np.ndarray) -> NumericValue:
        """
        Verify wavelet localization properties.
        
        From appendix_b_basis.tex Eq B.33:
        The wavelets must satisfy:
        ∫dx (1 + |x|²)|ψ(x)|² < ∞
        """
        try:
            # Load wavelet thresholds
            thresholds = self._load_wavelet_thresholds()
            
            # Compute localization measure
            x = np.linspace(-3, 3, 100)  # SACRED: grid range
            measure = np.sum((1 + x**2) * np.abs(transforms)**2)
            uncertainty = np.std(np.abs(transforms)**2)
            
            # Validate against threshold
            if measure > thresholds['localization']:
                raise ValidationError(f"Poor wavelet localization: {measure}")
                
            return NumericValue(measure, uncertainty)
            
        except Exception as e:
            raise ValidationError(f"Localization verification failed: {e}")

    def _verify_resolution(self, transforms: np.ndarray) -> NumericValue:
        """
        Verify resolution of unity property.
        
        From appendix_b_basis.tex Eq B.34:
        The wavelets must satisfy:
        ∫da/a ∫db |W_ψ[f](a,b)|² = ‖f‖²
        """
        try:
            # Load wavelet thresholds
            thresholds = self._load_wavelet_thresholds()
            
            # Compute resolution measure
            da = np.diff(self.alpha**np.arange(self.max_level))
            db = np.linspace(-3, 3, 100)[1] - np.linspace(-3, 3, 100)[0]  # SACRED
            
            measure = np.sum(np.abs(transforms)**2 * da[:, None] * db)
            uncertainty = np.std(np.abs(transforms)**2) * np.sqrt(len(da) * len(transforms[0]))
            
            # Validate against threshold
            if abs(measure - 1.0) > thresholds['resolution']:
                raise ValidationError(f"Resolution of unity violated: {measure}")
                
            return NumericValue(measure, uncertainty)
            
        except Exception as e:
            raise ValidationError(f"Resolution verification failed: {e}")

    def _verify_admissibility(self, transforms: np.ndarray) -> NumericValue:
        """
        Verify wavelet admissibility condition.
        
        From appendix_b_basis.tex Eq B.35:
        The wavelets must satisfy:
        C_ψ = ∫dω |ω|^{-1} |ψ̂(ω)|² < ∞
        """
        try:
            # Load wavelet thresholds
            thresholds = self._load_wavelet_thresholds()
            
            # Compute Fourier transform
            omega = np.fft.fftfreq(transforms.shape[1])
            psi_hat = np.fft.fft(transforms, axis=1)
            
            # Compute admissibility constant
            C_psi = np.sum(np.abs(psi_hat)**2 / np.abs(omega + 1e-10))
            uncertainty = np.std(np.abs(psi_hat)**2) / np.sqrt(len(omega))
            
            # Validate against threshold
            if C_psi > thresholds['admissibility']:
                raise ValidationError(f"Admissibility condition violated: {C_psi}")
                
            return NumericValue(C_psi, uncertainty)
            
        except Exception as e:
            raise ValidationError(f"Admissibility verification failed: {e}")

    def _validate_wavelet_analysis(self, results: Dict[str, NumericValue]) -> None:
        """
        Validate complete wavelet analysis results.
        
        From appendix_b_basis.tex Eq B.36-B.38:
        The wavelet analysis must satisfy:
        1. Proper localization
        2. Resolution of unity
        3. Admissibility
        4. Frame bounds
        """
        try:
            # Load validation data
            thresholds = self._load_wavelet_thresholds()
            
            # Check localization
            loc = results['localization'].value
            if loc > thresholds['max_localization']:
                raise ValidationError(f"Excessive delocalization: {loc}")
                
            # Verify resolution
            res = results['resolution'].value
            if abs(res - 1.0) > thresholds['resolution_tolerance']:
                raise ValidationError(f"Resolution of unity violated: {res}")
                
            # Check admissibility
            adm = results['admissibility'].value
            if not thresholds['min_admissibility'] < adm < thresholds['max_admissibility']:
                raise ValidationError(f"Admissibility bounds violated: {adm}")
                
        except Exception as e:
            raise ValidationError(f"Wavelet analysis validation failed: {e}")